requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
brats
redis==5.0.1
pydantic==2.5.0